requirements/base.txt
Django==3.2.9
django-environ==0.10.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==9.5.0
psycopg2-binary==2.9.5
requests==2.29.0
social-auth-app-django==4.0.0
